+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - War failing to start if the number of factions needed to be reduced to fit the number of available stars
 - Minions of the primary faction AI spawning as copies of the primary AI instead of themselves
+- Hummingbird Upgrade Tech not adding the appropriate unit type tag
 
 ## v5.42.0 - 2022-07-23